pyproject.toml
[tool.poetry]
name = "carrot-rcc"
version = "0.4.2"
description = "Camunda external task Robot Framework RCC client"
homepage = "https://github.com/datakurre/carrot-rcc"
authors = ["Asko Soukka <asko.soukka@iki.fi>"]
license = "MIT"
packages = [
    { include = "carrot_rcc.py" },
]
include = [ "carrot-rcc" ]
readme = "README.md"

[tool.poetry.scripts]
carrot-rcc= "carrot_rcc:main"

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
robotframework = "^4.1"
robotframework-lsp = "^0.20.0"
robotframework-robocop = "^1.8.1"
appdirs = "^1.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
